LRSD-7729 Enable spring boot caching with Caffeine, move out Jira calls so that they would be cacheable
